+++ lib/commands/plugin-test.sh
@@ -21,7 +21,7 @@ plugin_test_command() {
         fail_test "could not install $plugin_name from $plugin_url"
     fi
 
-    if ! (asdf plugin-list | grep "$plugin_name" > /dev/null); then
+    if ! (asdf plugin-list | grep "^$plugin_name$" > /dev/null); then
         fail_test "$plugin_name was not properly installed"
     fi