+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Fixed Bugs
 
+* Incorrect `grep` command caused version command to look at the wrong tool when reporting the version (#262)
+
 ## 0.4.1
 
 Features